Update AccountKey examples by using js-ext-core
